Front: ver meta, ver e editar carro

+++ src/findriver/back/service/goalService.js
@@ -1,4 +1,4 @@
-const supabase = require("./userService")
+const { supabase } = require("./userService")
 const jwt = require("jsonwebtoken")
 const env = require("dotenv")
 
@@ -10,19 +10,21 @@ const createGoal = async (userId, Goal) => {
             {
                 idUser: userId,
                 name: Goal.name,
-                valueCurrent: 0,
-                valueGoal: Goal.goal,
-                deadLine: Goal.deadLine,
+                valueCurrent: Goal.valueCurrent,
+                valueGoal: Goal.valueGoal,
+                deadline: Goal.deadline,
                 inicialDate: new Date().toISOString(),
             },
         ])
         .single();
+
+        if (error) throw error;
     } catch (error) {
         throw error
     }
 }
 
-const getGoalByUserID = async (userId) => {
+const getGoalByUserId = async (userId) => {
     try {
         const {data, error} = await supabase
             .from("Goal")
@@ -38,19 +40,17 @@ const getGoalByUserID = async (userId) => {
     }
 };
 
-const updateGoal = async (Goal, goalId) => {
+const updateGoal = async (Goal, userId) => {
     try { 
-        const {data, error} = await supabase
+        const {error} = await supabase
             .from("Goal")
             .update({
-                idUser: userId,
                 name: Goal.name,
                 valueCurrent: Goal.valueCurrent,
-                valueGoal: Goal.goal,
-                deadLine: Goal.deadLine,
-                inicialDate: Goal.inicialDate,
+                valueGoal: Goal.valueGoal,
+                deadline: Goal.deadline,
             })
-            .eq("id", goalId);
+            .eq("idUser", userId);
 
         if(error) throw error;
     } catch (error) {
@@ -58,15 +58,17 @@ const updateGoal = async (Goal, goalId) => {
     }
 };
 
-const deleteGoal = async (goalId) => {
+const deleteGoal = async (userId) => {
     try {
-        const {data, error} = await supabase
+        const {error} = await supabase
             .from("Goal")
             .delete()
-            .eq("id", goalId);
+            .eq("idUser", userId);
 
         if(error) throw error;
     } catch (error) {
         throw error;
     }
 };
+
+module.exports = {createGoal, getGoalByUserId, updateGoal, deleteGoal}
